ActionEditV2 refactor
